# platform/contracts/openapi/v1/wego-api.yaml
openapi: 3.1.1
info:
  title: Wego API
  version: 1.0.0
  description: >-
    Version 1 of the product-neutral Wego HTTP contract. WEGO-001 adds the
    first authenticated operations (login, logout, current identity);
    WEGO-002 adds the first business-domain operations (diving offerings and
    staff-created bookings).
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Current client instance
...
paths:
...
  /api/v1/tours-operator/staff/tours/{tourId}/media/upload:
    post:
      operationId: uploadToursOperatorTourMedia
      summary: Upload a private draft photo; requires media upload and tour manage
      description: JPEG/PNG only, at most 10 MiB and 24 million pixels. Server re-encodes without metadata. Appends up to 30 photos or replaces a mediaId only with its current reviewed revision. Never approves rights.
      tags: [ToursOperator]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: tourId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ToursOperatorMediaUploadRequest"
      responses:
        "200": { $ref: "#/components/responses/ToursOperatorMediaUploaded" }
        "400": { $ref: "#/components/responses/ToursOperatorMediaInvalid" }
        "401": { $ref: "#/components/responses/UnauthenticatedResponse" }
        "403": { $ref: "#/components/responses/ForbiddenResponse" }
        "404": { description: Tour or replacement media not found. }
        "409": { description: draft_changed; re-read before retrying. }
        "413": { description: Upload exceeds byte or multipart request limit. }
        "503": { description: storage_failed or upload_busy (Retry-After 3); do not blindly retry an append. }
  /api/v1/tours-operator/staff/categories/{category}/media/upload:
    post:
      operationId: uploadToursOperatorCategoryMedia
      summary: Upload a category draft cover with reviewed revision
      tags: [ToursOperator]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ToursOperatorMediaCategory"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/ToursOperatorMediaUploadRequest"
                - type: object
                  required: [revision]
      responses:
        "200": { $ref: "#/components/responses/ToursOperatorMediaUploaded" }
        "400": { $ref: "#/components/responses/ToursOperatorMediaInvalid" }
        "401": { $ref: "#/components/responses/UnauthenticatedResponse" }
        "403": { $ref: "#/components/responses/ForbiddenResponse" }
        "404": { description: Unknown category. }
        "409": { description: draft_changed. }
        "413": { description: Upload exceeds limit. }
        "503": { description: storage_failed or upload_busy. }
  /api/v1/tours-operator/categories/media:
    get:
      operationId: listToursOperatorPublicCategoryCovers
      summary: Approved category covers only, with requested alt or English fallback
      tags: [ToursOperator]
      security: []
      parameters:
        - name: locale
          in: query
          schema: { $ref: "#/components/schemas/ToursOperatorContentLocale" }
      responses:
        "200":
          description: No draft metadata or private storage keys.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ToursOperatorPublicCategoryCover" }
        "400": { description: unsupported_locale. }
  /api/v1/tours-operator/staff/categories/media:
    get:
      operationId: listToursOperatorCategoryMedia
      summary: List the five existing category covers and review revisions
      tags: [ToursOperator]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Staff draft/approved metadata.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ToursOperatorCategoryMedia" }
        "401": { $ref: "#/components/responses/UnauthenticatedResponse" }
        "403": { $ref: "#/components/responses/ForbiddenResponse" }
  /api/v1/tours-operator/staff/categories/{category}/media:
    get:
      operationId: getToursOperatorCategoryMedia
      summary: Read one category cover for editing
      tags: [ToursOperator]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ToursOperatorMediaCategory"
      responses:
        "200":
          description: Staff metadata including revision.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToursOperatorCategoryMedia" }
        "401": { $ref: "#/components/responses/UnauthenticatedResponse" }
        "403": { $ref: "#/components/responses/ForbiddenResponse" }
        "404": { description: Unknown category. }
  /api/v1/tours-operator/staff/categories/{category}/media/alt:
    put:
      operationId: updateToursOperatorCategoryMediaAlt
      summary: Edit alt texts with revision; changed text resets rights
      tags: [ToursOperator]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ToursOperatorMediaCategory"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ToursOperatorCategoryAltRequest" }
      responses:
        "204": { description: Updated. }
        "400": { $ref: "#/components/responses/ToursOperatorMediaInvalid" }
        "401": { $ref: "#/components/responses/UnauthenticatedResponse" }
        "403": { $ref: "#/components/responses/ForbiddenResponse" }
        "404": { description: Unknown category. }
        "409": { description: draft_changed. }
  /api/v1/tours-operator/staff/categories/{category}/media/approve:
    post:
      operationId: approveToursOperatorCategoryMedia
      summary: Approve exact reviewed category image and alt; requires content publish
      tags: [ToursOperator]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ToursOperatorMediaCategory"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ToursOperatorPublishRequest" }
      responses:
        "204": { description: Approved. }
        "400": { $ref: "#/components/responses/ToursOperatorMediaInvalid" }
        "401": { $ref: "#/components/responses/UnauthenticatedResponse" }
        "403": { $ref: "#/components/responses/ForbiddenResponse" }
        "404": { description: Unknown category. }
        "409": { description: draft_changed or no_asset. }
  /api/v1/tours-operator/staff/media/preview/{assetId}:
    get:
      operationId: previewToursOperatorMedia
      summary: Authorized private draft preview; never put bearer token in URL
      tags: [ToursOperator]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: assetId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: variant
          in: query
          schema: { type: string, enum: [base, w360, w768, w1024, w1440], default: base }
      responses:
        "200":
          description: Re-encoded JPEG/PNG; Cache-Control no-store, private.
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
        "401": { $ref: "#/components/responses/UnauthenticatedResponse" }
        "403": { $ref: "#/components/responses/ForbiddenResponse" }
        "404": { description: Asset or requested variant absent. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token issued by POST /api/v1/identity/login.
  responses:
    ToursOperatorMediaUploaded:
      description: Private asset registered; refresh staff content to read its path, dimensions and revision.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ToursOperatorMediaUploadResponse" }
    ToursOperatorMediaInvalid:
      description: Invalid image, alt or media selection.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ToursOperatorMediaError" }
...
    UnauthenticatedResponse:
      description: Missing, invalid, or expired session token.
      headers:
        WWW-Authenticate:
          description: Always `Bearer` — the authentication scheme this API accepts.
          schema:
            type: string
            enum:
              - Bearer
    ForbiddenResponse:
      description: Authenticated, but missing the permission this operation requires.
...
  parameters:
    ToursOperatorMediaCategory:
      name: category
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ToursOperatorCategory" }
...
  schemas:
    ToursOperatorMediaUploadResponse:
      type: object
      additionalProperties: false
      required: [assetId, mediaId]
      properties:
        assetId: { type: string, format: uuid }
        mediaId: { type: [string, "null"], format: uuid }
    ToursOperatorMediaError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { type: string }
    ToursOperatorMediaUploadRequest:
      type: object
      required: [file, requestId]
      properties:
        file: { type: string, format: binary }
        requestId:
          type: string
          format: uuid
          description: Stable per-attempt UUID. Replay with same owner, bytes and alt returns same linked asset; reuse with changed input is rejected.
        mediaId: { type: string, format: uuid }
        revision: { type: string, maxLength: 64 }
        altEn: { type: string, maxLength: 200 }
        altAr: { type: string, maxLength: 200 }
        altRu: { type: string, maxLength: 200 }
        altIt: { type: string, maxLength: 200 }
    ToursOperatorCategoryAltRequest:
      type: object
      required: [revision]
      properties:
        revision: { type: string, maxLength: 64 }
        altEn: { type: [string, "null"], maxLength: 200 }
        altAr: { type: [string, "null"], maxLength: 200 }
        altRu: { type: [string, "null"], maxLength: 200 }
        altIt: { type: [string, "null"], maxLength: 200 }
...
    ToursOperatorPublicCategoryCover:
      type: object
      additionalProperties: false
      required: [category, path, width, height, alt]
      properties:
        category: { $ref: "#/components/schemas/ToursOperatorCategory" }
        path: { type: string }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        alt: { type: string }
    ToursOperatorCategoryMedia:
      type: object
      required: [category, assetId, alt, rightsStatus, approvedAt, updatedAt, revision, path, width, height]
      properties:
        category: { $ref: "#/components/schemas/ToursOperatorCategory" }
        assetId: { type: [string, "null"], format: uuid }
        alt:
          type: object
          additionalProperties: { type: string, maxLength: 200 }
        rightsStatus: { type: string, enum: [DRAFT, APPROVED] }
        approvedAt: { type: [string, "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
        revision: { type: string }
        path: { type: [string, "null"] }
        width: { type: [integer, "null"], minimum: 1 }
        height: { type: [integer, "null"], minimum: 1 }
...
    ToursOperatorCategory:
      type: string
      enum:
        - DESERT
        - SEA
        - CULTURAL
        - SHOWS
        - TRANSFERS
...
    ToursOperatorPublishRequest:
      type: object
      additionalProperties: false
      required:
        - revision
      properties:
        revision:
          type: string
          minLength: 1
          description: Revision of the draft (or media item) the publisher reviewed.
    ToursOperatorContentLocale:
      type: string
      enum:
        - en
        - ar
        - ru
        - it
...